[ALPS07682698] Revert "ACK: Merge android 5.15.41 temporarily"
This reverts commit ee30f62d45c711758280981870607498b30209b8.

CR-Id: ALPS07682698
Feature: Kernel SI Operation

+++ drivers/gpu/drm/amd/amdgpu/amdgpu_bo_list.h
@@ -61,7 +61,7 @@ int amdgpu_bo_create_list_entry_array(struct drm_amdgpu_bo_list_in *in,
 int amdgpu_bo_list_create(struct amdgpu_device *adev,
 				 struct drm_file *filp,
 				 struct drm_amdgpu_bo_list_entry *info,
-				 size_t num_entries,
+				 unsigned num_entries,
 				 struct amdgpu_bo_list **list);
 
 static inline struct amdgpu_bo_list_entry *
